Added List, Main and About views

+++ GitTagApp/Entities/GitRepoTag.cs
@@ -2,10 +2,10 @@
 {
     public class GitRepoTag
     {
-        public int GitRepoId { get; set; }
+        public long GitRepoId { get; set; }
         public GitRepo GitRepo { get; set; }
         
-        public int TagId { get; set; }
+        public long TagId { get; set; }
         public Tag Tag { get; set; }
     }
 }
