+++ GitTagApp/Interfaces/IGenericRepository.cs
@@ -6,9 +6,9 @@ namespace GitTagApp.Interfaces
     public interface IGenericRepository<T> : IDisposable where T : class, IBaseEntity
     {
         IEnumerable<T> GetAll();
-        T GetById(int id);
+        T GetById(long id);
         void Create(T entity);
         void Update(T entity);
-        void Delete(int id);
+        void Delete(long id);
     }
 }
